+++ TeamProject/Poker/Models/Players/Bot.cs
@@ -12,7 +12,7 @@ namespace Poker.Players
     public class Bot : Character, IBot
     {
 
-        public Bot(string name, int chips, int power = 0) 
+        public Bot(string name, int chips, float power = 0) 
             :base(name, chips, power)
         {
             this.Hand = new List<ICard>();
